Limit audio bandwidth with BPF not calculated timing
Add a constant specifying the target audio bandwidth of the produced
morse signal, and enforce this bandwidth using stacked IIR bandpass
filters.

This replaces the old behavior of manually ramping the dit/dah volume
from/to zero at the beginning/end of the symbol. The
symbol-length-dependent ramping produced a fairly wide bandwidth,
especially at faster code speeds, and was difficult to tune.

Set the default bandwidth to 200 Hz, centered on the code audio
frequency.