Bugfix: Define new BPF for each letter() call
Defining the band pass filter globally prevents changing the tone HERTZ
constant at runtime (e.g. with a context manager), since the BPF will
still be defined with the default HERTZ as the center frequency.

Move the bpf generation into the letter() call where it'll be used to
allow for overriding HERTZ.